Add some input protection on the xcresult path
